Use logger in demo code
If a new user copies the current code on the homepage, they may be surprised to see no logs in their console when making a requests. While including the `logger` option makes the example more bloated, I think having the logs in and running is closer to a real world example.

An alternative here would be to remove all of the log lines but I don't think that is valuable.